feat: add role and permission schedule and tahfidz

+++ app/Http/Controllers/Admin/ScheduleController.php
@@ -2,12 +2,14 @@
 
 namespace App\Http\Controllers\Admin;
 
-use App\Http\Controllers\Controller;
-use App\Http\Requests\Admin\ScheduleRequest;
-use App\Models\Schedule;
+use Carbon\Carbon;
 use App\Models\School;
+use App\Models\Schedule;
 use Illuminate\Http\Request;
 use Yajra\DataTables\DataTables;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
+use App\Http\Requests\Admin\ScheduleRequest;
 
 class ScheduleController extends Controller
 {
@@ -16,14 +18,18 @@ class ScheduleController extends Controller
      */
     public function index()
     {
+        if (!Auth::user()->can('Manage Jadwal')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
+
         if (request()->ajax()) {
-            $data = Schedule::with('school')->orderBy('date', 'desc')->get();
+            $data = Schedule::with('school')->orderBy('date', 'desc');
             return DataTables::of($data)
                 ->editColumn('type', function ($data) {
                     return $data->type == Schedule::TYPE_ALL ? '<span class="badge badge-success">Semua</span>' : '<span class="badge badge-primary">Sekolah</span>';
                 })
                 ->editColumn('date', function ($data) {
-                    return $data->date->format('d F Y');
+                    return Carbon::parse($data->date)->locale('id')->translatedFormat('l, d F Y');
                 })
                 ->editColumn('school_id', function ($data) {
                     return $data->school_id ? '<span class="badge badge-primary">' . $data->school->name . '</span>' : '<span class="badge badge-secondary">Semua Sekolah</span>';
@@ -32,8 +38,8 @@ class ScheduleController extends Controller
                     $actionEdit = route('schedule.edit', $data->id);
                     $actionDelete = route('schedule.destroy', $data->id);
                     return "<div class='d-flex justify-content-center'>" .
-                        view('components.action.edit', ['action' => $actionEdit]) .
-                        view('components.action.delete', ['action' => $actionDelete, 'id' => $data->id]) .
+                        view('components.action.edit', ['action' => $actionEdit, 'name' => 'Jadwal']) . '&nbsp;' .
+                        view('components.action.delete', ['action' => $actionDelete, 'id' => $data->id, 'name' => 'Jadwal']) .
                         "</div>";
                 })
                 ->rawColumns(['action', 'school_id', 'type'])
@@ -47,6 +53,9 @@ class ScheduleController extends Controller
      */
     public function create()
     {
+        if (!Auth::user()->can('Create Jadwal')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $schools = School::orderBy('name')->get();
         return view('admins.schedule.create-edit', compact('schools'));
     }
@@ -56,6 +65,9 @@ class ScheduleController extends Controller
      */
     public function store(ScheduleRequest $request)
     {
+        if (!Auth::user()->can('Create Jadwal')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $data = $request->validated();
         Schedule::create($data);
         return redirect()->route('schedule.index')->with('success', 'Data berhasil ditambahkan');
@@ -74,6 +86,9 @@ class ScheduleController extends Controller
      */
     public function edit(Schedule $schedule)
     {
+        if (!Auth::user()->can('Edit Jadwal')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $schools = School::orderBy('name')->get();
         return view('admins.schedule.create-edit', compact('schedule', 'schools'));
     }
@@ -83,6 +98,9 @@ class ScheduleController extends Controller
      */
     public function update(ScheduleRequest $request, Schedule $schedule)
     {
+        if (!Auth::user()->can('Edit Jadwal')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $data = $request->validated();
         $schedule->update($data);
         return redirect()->route('schedule.index')->with('success', 'Data berhasil diubah');
@@ -93,6 +111,9 @@ class ScheduleController extends Controller
      */
     public function destroy(Schedule $schedule)
     {
+        if (!Auth::user()->can('Delete Jadwal')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $schedule->delete();
         return redirect()->route('schedule.index')->with('success', 'Data berhasil dihapus');
     }
